pyproject.toml
[project]
name = "practice-tools"
version = "0.1.0"
description = "Collection of music practice tools"
requires-python = ">=3.8"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "mysql-connector-python>=8.2.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
